refactor: move BriefCardComponent to `features/product/components`
