+++ src/app/features/product/components/brief-card/brief-card.component.html
